Added documentation to PrimaryAttributes and SecondaryAttributes

+++ Characters/Mage.cs
@@ -98,9 +98,9 @@ namespace RPG_Characters
         {
             if (Equipment[Slot.SLOT_WEAPON] != null)
             {
-                return (double)((double)Equipment[Slot.SLOT_WEAPON].GetDPS() + 1.0 + ((double)BasePrimaryAttributes.Intelligence + (double)TotalPrimaryAttributes.Intelligence) / 100);
+                return (double)((double)Equipment[Slot.SLOT_WEAPON].GetDPS() + 1.0 + (double)(BasePrimaryAttributes.Intelligence + TotalPrimaryAttributes.Intelligence) / 100);
             }
-            return 1 + ((double)(BasePrimaryAttributes.Intelligence + (double)TotalPrimaryAttributes.Intelligence) / 100);
+            return 1 + (double)((BasePrimaryAttributes.Intelligence + TotalPrimaryAttributes.Intelligence) / 100);
         }
     }
 }
